stateInMemory: absorb kvstore.js, change exports
